# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для тестирования стажёров Активбанка через Google Forms.",
        "title": "Activ GForms API",
        "contact": {
            "name": "Amin Muborakkadamov",
            "url": "https://github.com/AminVamar"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/attempts/{id}": {
            "get": {
                "description": "Все ответы и итоговый балл.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "попытки"
                ],
                "summary": "Попытка целиком",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.AttemptResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/attempts/{id}/rescore": {
            "post": {
                "description": "Ручные баллы не меняются.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "попытки"
                ],
                "summary": "Пересчёт автоматических баллов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.RescoreResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "источник тестов недоступен",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/interns": {
            "get": {
                "description": "Поиск по ФИО или телефону.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "стажёры"
                ],
                "summary": "Стажёры",
                "parameters": [
                    {
                        "type": "string",
                        "description": "часть ФИО или телефона",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "фильтр по филиалу",
                        "name": "branch_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "лимит (до 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.ListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.InternListItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/interns/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "стажёры"
                ],
                "summary": "Стажёр и его попытки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "фильтр по тесту",
                        "name": "test_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "graded или needs_grading",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.InternCardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Answer": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "attempt_id": {
                    "type": "integer"
                },
                "correct_answers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "graded_at": {
                    "type": "string"
                },
                "graded_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "domain.Attempt": {
            "type": "object",
            "properties": {
                "auto_submitted": {
                    "type": "boolean"
                },
                "branch_code": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "integer"
                },
                "branch_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "form_session_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "intern_id": {
                    "type": "integer"
                },
                "intern_name": {
                    "type": "string"
                },
                "intern_phone": {
                    "type": "string"
                },
                "needs_grading": {
                    "type": "boolean"
                },
                "pending_count": {
                    "type": "integer"
                },
                "response_id": {
                    "type": "string"
                },
                "scored": {
                    "type": "integer"
                },
                "submitted_at": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "test_title": {
                    "type": "string"
                },
                "total_questions": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Intern": {
            "type": "object",
            "properties": {
                "branch_code": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "integer"
                },
                "branch_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.InternListItem": {
            "type": "object",
            "properties": {
                "attempts_count": {
                    "type": "integer"
                },
                "branch_code": {
                    "type": "string"
                },
                "branch_id": {
                    "type": "integer"
                },
                "branch_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "needs_grading": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "http.AttemptResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Answer"
                    }
                },
                "attempt": {
                    "$ref": "#/definitions/domain.Attempt"
                },
                "score": {
                    "type": "integer"
                },
                "source_available": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "http.InternCardResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Attempt"
                    }
                },
                "attempts_total": {
                    "type": "integer"
                },
                "intern": {
                    "$ref": "#/definitions/domain.Intern"
                }
            }
        },
        "http.ListResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "http.RescoreResponse": {
            "type": "object",
            "properties": {
                "attempt_id": {
                    "type": "integer"
                },
                "changed": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}